requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.2
aniso8601==9.0.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
Flask==2.3.2
Flask-RESTful==0.3.10
frozenlist==1.3.3
future==0.18.2
heyoo==0.0.9
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.15.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
pip==23.1.2
pymongo==4.4.0
pyngrok==6.0.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-toolbelt==1.0.0
setuptools==58.0.4
six==1.15.0
tqdm==4.65.0
typing==3.7.4.3
urllib3==2.0.3
Werkzeug==2.3.6
wheel==0.37.0
yarl==1.9.2
zipp==3.15.0
